# swagger.json
{"openapi":"3.0.0","paths":{"/api-docs":{"get":{"operationId":"getApiDocs","summary":"","description":"","parameters":[],"responses":{"200":{"description":""}}}},"/event/{id}":{"get":{"operationId":"getEvent","summary":"","description":"---------------------------------- event ---------------------------------","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/EventEntity"}}}}},"tags":["event"]},"put":{"operationId":"updateEvent","summary":"","description":"","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateEventDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/EventEntity"}}}}},"tags":["event"]},"delete":{"operationId":"removeEvent","summary":"","description":"","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/EventEntity"}}}}},"tags":["event"]}},"/event":{"get":{"operationId":"getEvents","summary":"","description":"根据时间范围或给定id列表获取事件，将二者的并集返回","parameters":[{"name":"range","required":false,"in":"query","schema":{"type":"string"}},{"name":"ids","required":false,"in":"query","schema":{"type":"array","items":{"type":"number"}}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/EventEntity"}}}}}},"tags":["event"]},"post":{"operationId":"createEvent","summary":"","description":"","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateEventDto"}}}},"responses":{"201":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/EventEntity"}}}}},"tags":["event"]}},"/event/{id}/content":{"get":{"operationId":"getContent","summary":"","description":"--------------------------------- content --------------------------------","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":""}},"tags":["event"]},"post":{"operationId":"createContent","summary":"","description":"","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateContentDto"}}}},"responses":{"201":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/EventContentEntity"}}}}},"tags":["event"]},"put":{"operationId":"updateContent","summary":"","description":"","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateContentDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/EventContentEntity"}}}}},"tags":["event"]}},"/event/{id}/todo":{"get":{"operationId":"getTodos","summary":"","description":"---------------------------------- todo ----------------------------------","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":""}},"tags":["event"]},"post":{"operationId":"createTodo","summary":"","description":"","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"201":{"description":""}},"tags":["event"]},"put":{"operationId":"updateTodo","summary":"","description":"","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":""}},"tags":["event"]},"delete":{"operationId":"removeTodo","summary":"","description":"","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":""}},"tags":["event"]}},"/user":{"get":{"operationId":"getUser","summary":"","description":"","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserEntity"}}}}},"tags":["user"]},"delete":{"operationId":"deleteUser","summary":"","description":"","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserEntity"}}}}},"tags":["user"]}},"/user/login":{"post":{"operationId":"login","summary":"","description":"","parameters":[],"responses":{"201":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserLoginRespDto"}}}}},"tags":["user"]}},"/user/register":{"post":{"operationId":"register","summary":"","description":"","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateUserDto"}}}},"responses":{"201":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserEntity"}}}}},"tags":["user"]}},"/project":{"post":{"operationId":"createProject","summary":"","description":"创建新项目","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateProjectDto"}}}},"responses":{"201":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProjectEntity"}}}}},"tags":["project"]},"get":{"operationId":"getProject","summary":"","description":"获取项目信息","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProjectEntity"}}}}},"tags":["project"]},"put":{"operationId":"updateProject","summary":"","description":"更新项目信息","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateProjectDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProjectEntity"}}}}},"tags":["project"]},"delete":{"operationId":"removeProject","summary":"","description":"删除项目","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProjectEntity"}}}}},"tags":["project"]}},"/project/workspace":{"get":{"operationId":"getWorkspaceInfo","summary":"","description":"获取工作区信息","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/WorkspaceEntity"}}}}},"tags":["project"]},"put":{"operationId":"updateWorkspaceInfo","summary":"","description":"更新工作区信息","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateWorkspaceDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/WorkspaceEntity"}}}}},"tags":["project"]}},"/project/settings":{"get":{"operationId":"getProjectSettings","summary":"","description":"获取项目设置","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"object"}}}}},"tags":["project"]},"put":{"operationId":"updateProjectSettings","summary":"","description":"更新项目设置","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateSettingsDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SettingsEntity"}}}}},"tags":["project"]}},"/character":{"post":{"operationId":"create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateCharacterDto"}}}},"responses":{"201":{"description":""}}},"get":{"operationId":"findAll","parameters":[],"responses":{"200":{"description":""}}}},"/character/{id}":{"get":{"operationId":"findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}}},"patch":{"operationId":"update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateCharacterDto"}}}},"responses":{"200":{"description":""}}},"delete":{"operationId":"remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}}}},"/scence":{"post":{"operationId":"create","summary":"","description":"","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateScenceDto"}}}},"responses":{"201":{"description":"","content":{"application/json":{"schema":{"type":"string"}}}}}},"get":{"operationId":"findAll","summary":"","description":"","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"string"}}}}}}},"/scence/{id}":{"get":{"operationId":"findOne","summary":"","description":"","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"string"}}}}}},"patch":{"operationId":"update","summary":"","description":"","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateScenceDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"string"}}}}}},"delete":{"operationId":"remove","summary":"","description":"","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"string"}}}}}}},"/backdrop":{"post":{"operationId":"create","summary":"","description":"","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateBackdropDto"}}}},"responses":{"201":{"description":"","content":{"application/json":{"schema":{"type":"string"}}}}}},"get":{"operationId":"findAll","summary":"","description":"","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"string"}}}}}}},"/backdrop/{id}":{"get":{"operationId":"findOne","summary":"","description":"","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"string"}}}}}},"patch":{"operationId":"update","summary":"","description":"","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateBackdropDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"string"}}}}}},"delete":{"operationId":"remove","summary":"","description":"","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"string"}}}}}}}},"info":{"title":"","description":"","version":"1.0.0","contact":{}},"tags":[],"servers":[],"components":{"securitySchemes":{"bearer":{"scheme":"bearer","bearerFormat":"JWT","type":"http"}},"schemas":{"EventEntity":{"type":"object","properties":{"type":{"type":"string","enum":["ATOM","COLLECTION"]},"range":{"type":"string"},"id":{"type":"number"},"createdAt":{"format":"date-time","type":"string"},"updatedAt":{"format":"date-time","type":"string"},"deletedAt":{"format":"date-time","type":"string","nullable":true},"contentId":{"type":"number","nullable":true},"projectId":{"type":"number"},"name":{"type":"string"},"description":{"type":"string","nullable":true},"color":{"type":"string"}},"required":["type","range","id","createdAt","updatedAt","deletedAt","contentId","projectId","name","description","color"]},"CreateEventDto":{"type":"object","properties":{"range":{"type":"string"},"name":{"type":"string"},"description":{"type":"string","nullable":true},"color":{"type":"object","default":"#93c5fd"},"type":{"type":"object"},"graphId":{"type":"string"}},"required":["range","name","description","color","type"]},"UpdateEventDto":{"type":"object","properties":{"range":{"type":"string"},"name":{"type":"string"},"description":{"type":"string","nullable":true},"color":{"type":"object","default":"#93c5fd"},"type":{"type":"object"}}},"CreateContentDto":{"type":"object","properties":{"content":{"type":"string"}},"required":["content"]},"EventContentEntity":{"type":"object","properties":{"id":{"type":"number"},"updatedAt":{"format":"date-time","type":"string"},"eventId":{"type":"number"},"content":{"type":"string"}},"required":["id","updatedAt","eventId","content"]},"UpdateContentDto":{"type":"object","properties":{"content":{"type":"string"}}},"UserEntity":{"type":"object","properties":{"id":{"type":"number"},"username":{"type":"string"},"phone":{"type":"string","nullable":true},"email":{"type":"string","nullable":true}},"required":["id","username","phone","email"]},"UserLoginRespDto":{"type":"object","properties":{"access_token":{"type":"string"}},"required":["access_token"]},"CreateUserDto":{"type":"object","properties":{"username":{"type":"string"},"password":{"type":"string"},"phone":{"type":"string"},"email":{"type":"string"}},"required":["username","password"]},"CreateProjectDto":{"type":"object","properties":{"name":{"type":"string"},"description":{"type":"string"}},"required":["name","description"]},"SettingsEntity":{"type":"object","properties":{"darkMode":{"type":"boolean"}},"required":["darkMode"]},"LayoutEntity":{"type":"object","properties":{"id":{"type":"string"},"position":{"type":"array","items":{"type":"number"}}},"required":["id","position"]},"WorkspaceEntity":{"type":"object","properties":{"origin":{"type":"string"},"layout":{"type":"array","items":{"$ref":"#/components/schemas/LayoutEntity"}}},"required":["origin","layout"]},"ProjectEntity":{"type":"object","properties":{"settings":{"$ref":"#/components/schemas/SettingsEntity"},"workspace":{"$ref":"#/components/schemas/WorkspaceEntity"},"id":{"type":"number"},"name":{"type":"string"},"description":{"type":"string","nullable":true},"createdAt":{"format":"date-time","type":"string"},"updatedAt":{"format":"date-time","type":"string"},"deletedAt":{"format":"date-time","type":"string","nullable":true},"userId":{"type":"number"}},"required":["settings","workspace","id","name","description","createdAt","updatedAt","deletedAt","userId"]},"UpdateProjectDto":{"type":"object","properties":{"name":{"type":"string"},"description":{"type":"string"}}},"CreateLayoutDto":{"type":"object","properties":{"id":{"type":"string"},"position":{"type":"array","items":{"type":"number"}}},"required":["id","position"]},"UpdateWorkspaceDto":{"type":"object","properties":{"origin":{"type":"string"},"layout":{"type":"array","items":{"$ref":"#/components/schemas/CreateLayoutDto"}}}},"UpdateSettingsDto":{"type":"object","properties":{"darkMode":{"type":"boolean"}}},"CreateCharacterDto":{"type":"object","properties":{"name":{"type":"string"},"alias":{"type":"array","items":{"type":"string"}},"description":{"type":"string"},"unit":{"type":"number"},"start":{"format":"date-time","type":"string"},"end":{"format":"date-time","type":"string"}},"required":["name","alias","description","unit","start","end"]},"UpdateCharacterDto":{"type":"object","properties":{}},"CreateScenceDto":{"type":"object","properties":{}},"UpdateScenceDto":{"type":"object","properties":{}},"CreateBackdropDto":{"type":"object","properties":{}},"UpdateBackdropDto":{"type":"object","properties":{}}}}}